Remove stopwords file from Perl::Critic config
This will already be handled by a dzil test.  If we use Perl::Critic to
test spelling it won't have some of the default stopwords that the dzil
plugin uses and we'll just end up having to copy them over anyway.

Ideally the plugin might add the default stopwords to a file if it
doesn't have them, but that's a problem for another day.